Add (safe) uint32_t cast to fix Win64 build.
TBR=kwiberg@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/36539004

git-svn-id: http://webrtc.googlecode.com/svn/trunk@7916 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ webrtc/common_audio/wav_file.cc
@@ -68,7 +68,7 @@ size_t WavReader::ReadSamples(size_t num_samples, int16_t* samples) {
   // If we didn't read what was requested, ensure we've reached the EOF.
   CHECK(read == num_samples || feof(file_handle_));
   CHECK_LE(read, num_samples_remaining_);
-  num_samples_remaining_ -= read;
+  num_samples_remaining_ -= static_cast<uint32_t>(read);
   return read;
 }
